xaml language service inherits from xml language service.
