+++ AvalonStudio/AvalonStudio.Languages.Xaml/XmlLanguageService.cs
@@ -1,0 +1,141 @@
+﻿using Avalonia.Ide.CompletionEngine;
+using Avalonia.Ide.CompletionEngine.AssemblyMetadata;
+using Avalonia.Ide.CompletionEngine.SrmMetadataProvider;
+using AvaloniaEdit.Document;
+using AvaloniaEdit.Indentation;
+using AvalonStudio.Editor;
+using AvalonStudio.Extensibility.Languages.CompletionAssistance;
+using AvalonStudio.Projects;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AvalonStudio.Languages.Xaml
+{
+    class XmlLanguageService : ILanguageService
+    {
+        private static List<ICodeEditorInputHelper> s_InputHelpers = new List<ICodeEditorInputHelper>
+        {
+            new CompleteCloseTagCodeEditorHelper(),
+            new TerminateElementCodeEditorHelper(),
+            new InsertQuotesForPropertyValueCodeEditorHelper(),
+            new InsertExtraNewLineBetweenAttributesOnEnterCodeInputHelper()
+        };
+
+        public IIndentationStrategy IndentationStrategy { get; } = new XamlIndentationStrategy();
+
+        public virtual string Title => "XML";
+
+        public virtual string LanguageId => "xml";
+
+        public Type BaseTemplateType => null;
+
+        public IEnumerable<ICodeEditorInputHelper> InputHelpers => s_InputHelpers;
+
+        public IDictionary<string, Func<string, string>> SnippetCodeGenerators => new Dictionary<string, Func<string, string>>();
+
+        public IDictionary<string, Func<int, int, int, string>> SnippetDynamicVariables => new Dictionary<string, Func<int, int, int, string>>();
+
+        public IEnumerable<char> IntellisenseSearchCharacters => new[]
+        {
+            '(', ')', '.', ':', '-', '<', '>', '[', ']', ';', '"', '#', ','
+        };
+
+        public IEnumerable<char> IntellisenseCompleteCharacters => new[]
+        {
+            ',', '.', ':', ';', '-', ' ', '(', ')', '[', ']', '<', '>', '=', '+', '*', '/', '%', '|', '&', '!', '^'
+        };
+
+        public IObservable<TextSegmentCollection<Diagnostic>> Diagnostics => null;
+
+        public void Activation()
+        {
+        }
+
+        public void BeforeActivation()
+        {
+        }
+
+        public virtual bool CanHandle(ISourceFile file)
+        {
+            var result = false;
+
+            switch (Path.GetExtension(file.Location))
+            {
+                case ".xml":
+                    result = true;
+                    break;
+            }
+
+            return result;
+        }
+
+        public virtual bool CanTriggerIntellisense(char currentChar, char previousChar)
+        {
+            bool result = false;
+
+            if (currentChar == '<' || currentChar == ' ' || currentChar == '.')
+            {
+                return true;
+            }
+
+            return result;
+        }
+
+        public virtual Task<CodeCompletionResults> CodeCompleteAtAsync(ISourceFile sourceFile, int index, int line, int column, List<UnsavedFile> unsavedFiles, char lastChar, string filter = "")
+        {
+            return Task.FromResult<CodeCompletionResults>(null);
+        }
+
+        public int Comment(ISourceFile file, TextDocument textDocument, int firstLine, int endLine, int caret = -1, bool format = true)
+        {
+            return caret;
+        }
+
+        public int Format(ISourceFile file, TextDocument textDocument, uint offset, uint length, int cursor)
+        {
+            return cursor;
+        }
+
+        public Task<Symbol> GetSymbolAsync(ISourceFile file, List<UnsavedFile> unsavedFiles, int offset)
+        {
+            return Task.FromResult<Symbol>(null);
+        }
+
+        public Task<List<Symbol>> GetSymbolsAsync(ISourceFile file, List<UnsavedFile> unsavedFiles, string name)
+        {
+            return Task.FromResult<List<Symbol>>(null);
+        }
+
+        public bool IsValidIdentifierCharacter(char data)
+        {
+            return char.IsLetterOrDigit(data);
+        }
+
+        public virtual void RegisterSourceFile(AvaloniaEdit.TextEditor editor, ISourceFile file, TextDocument textDocument)
+        {
+        }
+
+        public virtual void UnregisterSourceFile(AvaloniaEdit.TextEditor editor, ISourceFile file)
+        {
+
+        }
+
+        public Task<CodeAnalysisResults> RunCodeAnalysisAsync(ISourceFile file, TextDocument textDocument, List<UnsavedFile> unsavedFiles, Func<bool> interruptRequested)
+        {
+            return Task.FromResult(new CodeAnalysisResults());
+        }
+
+        public Task<SignatureHelp> SignatureHelp(ISourceFile file, UnsavedFile buffer, List<UnsavedFile> unsavedFiles, int line, int column, int offset, string methodName)
+        {
+            return Task.FromResult<SignatureHelp>(null);
+        }
+
+        public int UnComment(ISourceFile file, TextDocument textDocument, int firstLine, int endLine, int caret = -1, bool format = true)
+        {
+            return caret;
+        }
+    }
+}
